misc(scripts): support TracingStartedInBrowser in scripts
