+++ lighthouse-core/gather/computed/trace-of-tab.js
@@ -18,6 +18,7 @@
 
 const ComputedArtifact = require('./computed-artifact');
 const log = require('lighthouse-logger');
+const TracingProcessor = require('../../lib/traces/tracing-processor');
 const LHError = require('../../lib/errors');
 const Sentry = require('../../lib/sentry');
 
@@ -51,31 +52,8 @@ class TraceOfTab extends ComputedArtifact {
       // @ts-ignore - stableSort added to Array by WebInspector.
       .stableSort((event0, event1) => event0.ts - event1.ts);
 
-    // Find out the inspected page frame.
-    /** @type {LH.TraceEvent|undefined} */
-    let startedInPageEvt;
-    const startedInBrowserEvt = keyEvents.find(e => e.name === 'TracingStartedInBrowser');
-    if (startedInBrowserEvt && startedInBrowserEvt.args.data &&
-        startedInBrowserEvt.args.data.frames) {
-      const mainFrame = startedInBrowserEvt.args.data.frames.find(frame => !frame.parent);
-      const pid = mainFrame && mainFrame.processId;
-      const threadNameEvt = keyEvents.find(e => e.pid === pid && e.ph === 'M' &&
-        e.cat === '__metadata' && e.name === 'thread_name' && e.args.name === 'CrRendererMain');
-      startedInPageEvt = mainFrame && threadNameEvt ?
-        Object.assign({}, startedInBrowserEvt, {
-          pid, tid: threadNameEvt.tid, name: 'TracingStartedInPage',
-          args: {data: {page: mainFrame.frame}}}) :
-        undefined;
-    }
-    // Support legacy browser versions that do not emit TracingStartedInBrowser event.
-    if (!startedInPageEvt) {
-      // The first TracingStartedInPage in the trace is definitely our renderer thread of interest
-      // Beware: the tracingStartedInPage event can appear slightly after a navigationStart
-      startedInPageEvt = keyEvents.find(e => e.name === 'TracingStartedInPage');
-    }
-    if (!startedInPageEvt) throw new LHError(LHError.errors.NO_TRACING_STARTED);
-    // @ts-ignore - property chain exists for 'TracingStartedInPage' event.
-    const frameId = startedInPageEvt.args.data.page;
+    // Find the inspected frame
+    const {startedInPageEvt, frameId} = TracingProcessor.findTracingStartedEvt(keyEvents);
 
     // Filter to just events matching the frame ID for sanity
     const frameEvents = keyEvents.filter(e => e.args.frame === frameId);
